Split dashboard display settings into focused tabs
