+++ frontend/src/views/Settings.vue
@@ -209,7 +209,7 @@ const settingsDestinations = [
   {
     id: "dashboard-layout",
     label: "Layout and calendar display",
-    path: "Dashboard / Layout & Calendar",
+    path: "Dashboard / Layout",
     category: "dashboard",
     tabKey: "settings_tab_dashboard",
     tab: "layout",
@@ -222,6 +222,31 @@ const settingsDestinations = [
       "time format",
       "meal plan",
     ],
+  },
+  {
+    id: "dashboard-calendar",
+    label: "Calendar display",
+    path: "Dashboard / Calendar Display",
+    category: "dashboard",
+    tabKey: "settings_tab_dashboard",
+    tab: "calendar",
+    keywords: [
+      "calendar",
+      "week",
+      "time format",
+      "weekend",
+      "red days",
+      "events",
+    ],
+  },
+  {
+    id: "dashboard-plugin-display",
+    label: "Plugin display",
+    path: "Dashboard / Plugin Display",
+    category: "dashboard",
+    tabKey: "settings_tab_dashboard",
+    tab: "plugin-display",
+    keywords: ["plugin", "display", "meal plan", "mealie", "card size"],
   },
   {
     id: "dashboard-ui",
